pOINTERS

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -9,11 +9,12 @@
 
 void _puts(char *str)
 {
-	long unsigned int i;
+	unsigned int i;
 
 	for (i = 0; i <= strlen(str); i++)
 	{
 		_putchar(str[i]);
+
 	}
 	_putchar('\n');
